ai-dashboard/requirements.txt
Flask==2.3.3
openai==0.28.1
python-dotenv==1.0.0
requests==2.31.0
plotly==5.17.0
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.7.1
matplotlib==3.8.2
seaborn==0.13.0
pillow==10.1.0
transformers==4.35.2
torch==2.1.2
sentence-transformers==2.2.2
langchain==0.0.350
langchain-community==0.0.10
chromadb==0.4.18
python-docx==0.8.11
PyPDF2==3.0.1 